+++ crates/teloxide-core/src/payloads/send_message.rs
@@ -3,9 +3,9 @@
 use serde::Serialize;
 
 use crate::types::{
-    BusinessConnectionId, CallbackQueryId, EffectId, LinkPreviewOptions, Message, MessageEntity,
-    ParseMode, Recipient, ReplyMarkup, ReplyParameters, SuggestedPostParameters, ThreadId, TopicId,
-    UserId,
+    BusinessConnectionId, EffectId, EphemeralMessageParameters, LinkPreviewOptions, Message,
+    MessageEntity, ParseMode, Recipient, ReplyMarkup, ReplyParameters, SuggestedPostParameters,
+    ThreadId, TopicId,
 };
 
 impl_payload! {
@@ -27,10 +27,8 @@ impl_payload! {
             pub message_thread_id: ThreadId,
             /// Identifier of the direct messages topic to which the message will be sent; required if the message is sent to a direct messages chat
             pub direct_messages_topic_id: TopicId,
-            /// For outgoing ephemeral messages, unique identifier of the user who will receive the message; for group and supergroup chats only
-            pub receiver_user_id: UserId,
-            /// For outgoing ephemeral messages, identifier of the callback query which triggered the message, if any
-            pub callback_query_id: CallbackQueryId,
+            /// A JSON-serialized object containing the parameters of the ephemeral message to send
+            pub ephemeral_message_parameters: EphemeralMessageParameters,
             /// Mode for parsing entities in the message text. See [formatting options] for more details.
             ///
             /// [formatting options]: https://core.telegram.org/bots/api#formatting-options